+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.9.0](https://github.com/it-objects/terraform-aws-terra3/compare/v1.8.0...v1.9.0) (2023-04-27)
+
+
+### Features
+
+* add alias domain ([#32](https://github.com/it-objects/terraform-aws-terra3/issues/32)) ([3a71814](https://github.com/it-objects/terraform-aws-terra3/commit/3a71814f56d67a87b08aedf4884db63388b9deba))
+
+## [1.8.0](https://github.com/it-objects/terraform-aws-terra3/compare/v1.7.0...v1.8.0) (2023-04-27)
+
+
+### Features
+
+* enabled S3 access control lists (ACLs) for S3 buckets ([#31](https://github.com/it-objects/terraform-aws-terra3/issues/31)) ([9201a3f](https://github.com/it-objects/terraform-aws-terra3/commit/9201a3fba70a2e5824a877249d0cd60a19e4464f))
+
 ## [1.7.0](https://github.com/it-objects/terraform-aws-terra3/compare/v1.6.0...v1.7.0) (2023-04-25)